+++ tests/cases/convert_libreoffice/main.sh
@@ -8,6 +8,11 @@ xargs rm -rvf < .gitignore
 export SOURCE_DATE_EPOCH="315532800"
 stepup -w 1 plan.py & # > current_stdout.txt &
 
+# Wait for the director and get its socket.
+export STEPUP_DIRECTOR_SOCKET=$(
+  python -c "import stepup.core.director; print(stepup.core.director.get_socket())"
+)
+
 # Get the graph after completion of the pending steps.
 python3 - << EOD
 from stepup.core.interact import *
@@ -27,7 +32,7 @@ write_manifest("reproducibility_manifest_skip.txt", ["slide.pdf", "slide1.pdf"])
 EOD
 
 # Wait for background processes, if any.
-wait $(jobs -p)
+wait
 
 # Check files that are expected to be present and/or missing.
 [[ -f plan.py ]] || exit -1