+++ tests/cases/pdflatex_bibtex/main.sh
@@ -9,6 +9,11 @@ export SOURCE_DATE_EPOCH="315532800"
 export LATEX_MAIN="paper"
 stepup -w 1 plan.py & # > current_stdout.txt &
 
+# Wait for the director and get its socket.
+export STEPUP_DIRECTOR_SOCKET=$(
+  python -c "import stepup.core.director; print(stepup.core.director.get_socket())"
+)
+
 # Get the graph after completion of the pending steps.
 python3 - << EOD
 from stepup.core.interact import *
@@ -32,7 +37,7 @@ write_manifest("reproducibility_bbl_manifest.txt", ["paper.bbl", "paper1.bbl"])
 EOD
 
 # Wait for background processes, if any.
-wait $(jobs -p)
+wait
 
 # Check files that are expected to be present and/or missing.
 [[ -f plan.py ]] || exit -1