+++ docs/changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Nicer fix for concurrent Inkscape SVG to PDF or PNG conversion
   (with `SELF_CALL=x`).
   See: https://gitlab.com/inkscape/inkscape/-/issues/4716
+- Make unit tests work with stepup-core 1.2.2.
 
 
 ## [1.1.1] - 2024-05-07